VariantLoadConfiguration no longer necessary, tests moved to VariantConfigurationTest
